introduce get_value and get_mod instead of from
